Fixed up the logout button

+++ app/src/main/java/com/codepath/apps/MainCode/models/Tweet.java
@@ -18,7 +18,7 @@ import java.util.Locale;
 public class Tweet {
     public String body;
     public String createdAt;
-    public int tweetID;
+    public long tweetID;
     public User user;
     public Entities entities;
     public String imageUrl;
@@ -30,10 +30,11 @@ public class Tweet {
 
     public Tweet() {}
 
+    // Build the tweet and gather data to fill it using the jsonObject throw the api request
     public static Tweet fromJson(JSONObject jsonObject) throws JSONException {
         Tweet tweet = new Tweet();
         tweet.body = jsonObject.getString("full_text");
-        tweet.tweetID = jsonObject.getInt("id");
+        tweet.tweetID = jsonObject.getLong("id");
         tweet.createdAt = jsonObject.getString("created_at");
         tweet.user = User.fromJson(jsonObject.getJSONObject("user"));
         if(!jsonObject.isNull("extended_entities")) {
